stress-udp-flood: cast integers to double for double division

+++ stress-udp-flood.c
@@ -170,11 +170,11 @@ static int OPTIMIZE3 stress_udp_flood(stress_args_t *args)
 	rate = (duration > 0.0) ? (bytes / duration) / (double)MB : 0.0;
 	stress_metrics_set(args, 0, "MB per sec sendto rate",
 		rate, STRESS_METRIC_HARMONIC_MEAN);
-	rate = (duration > 0.0) ? (stress_bogo_get(args) / duration) : 0.0;
+	rate = (duration > 0.0) ? ((double)stress_bogo_get(args) / duration) : 0.0;
 	stress_metrics_set(args, 1, "sendto calls per sec",
 		rate, STRESS_METRIC_HARMONIC_MEAN);
 	total_count = stress_bogo_get(args) + sendto_failed;
-	rate = (total_count > 0) ? ((total_count - sendto_failed) / total_count) * 100.0 : 0.0;
+	rate = (total_count > 0) ? ((double)(total_count - sendto_failed) / (double)total_count) * 100.0 : 0.0;
 	stress_metrics_set(args, 2, "% sendto calls succeeded",
 		rate, STRESS_METRIC_HARMONIC_MEAN);
